LoCo-SwiftUI now uses a more direct approach to updates.
Changes are propagated directly to the Store. We never update the Contacts array directly, we just let it update when the store changes and signals the change. This approach simplifies the merge logic. We no longer need to track what has changed since the last fetch. The UI is basically a direct mirror of the data.